feat(#295) :: repositroy 자료형 변경

+++ src/main/java/team/jeonghokim/daedongyeojido/domain/onepager/domain/repository/SubmitOnePagerRepository.java
@@ -3,5 +3,5 @@ package team.jeonghokim.daedongyeojido.domain.onepager.domain.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.jeonghokim.daedongyeojido.domain.onepager.domain.SubmitOnePager;
 
-public interface SubmitOnePagerRepository extends JpaRepository<SubmitOnePager, Integer> {
+public interface SubmitOnePagerRepository extends JpaRepository<SubmitOnePager, Long> {
 }
